+++ src/NumberNode.cpp
@@ -1,7 +1,7 @@
 #include "NumberNode.hpp"
 #include "NodeVisitor.hpp"
 
-NumberNode::NumberNode(int value): value(value) {}
+NumberNode::NumberNode(double value): value(value) {}
 
 void NumberNode::accept(NodeVisitor& visitor) {
 	visitor.visit(*this);
